refactor: most functionality moved to custom hooks
