+++ seankwarren/flowchart-calculator/imports/ui/components/mainPage/FlowchartCalculator.jsx
@@ -1,122 +1,54 @@
 import React, { useState, useEffect } from 'react';
-import Switch from 'react-switch';
 import 'reactflow/dist/style.css';
 import SplitPane from 'split-pane-react';
 import 'split-pane-react/esm/themes/default.css';
-import createNode from '../../utils/createNode';
 import startingNode from '../../utils/startingNode';
 import FlowchartCanvas from './FlowchartCanvas';
 import JSONViewer from './JSONViewer';
 import NodeButtons from './NodeButtons';
+import useLocalFlowData from '../../hooks/useLocalFlowData';
+import useResizeable from '../../hooks/useResizeable';
+import useRemoteFlowData from '../../hooks/useRemoteFlowData';
+import useDraggable from '../../hooks/useDraggable';
+import useToggleable from '../../hooks/useToggleable';
+import DarkModeSwitch from '../reusable/DarkModeSwitch';
 
 const FlowchartCalculator = () => {
     // state for nodes and edges
-    const [nodes, setNodes] = useState([startingNode]);
-    const [edges, setEdges] = useState([]);
-
-    // state for fetching saved flows from db
-    const [fetchedFlows, setFetchedFlows] = useState([]);
-
-    // state for user input flow name
-    const [flowName, setFlowName] = useState('');
-
-
-    const saveFlow = () => {
-        // Check if a flow with the same name already exists
-        const flowWithNameExists = fetchedFlows.some(flow => flow.name === flowName);
-        
-        if (flowWithNameExists) {
-            alert('A flow with this name already exists. Please choose a different name.');
-        } else {
-            Meteor.call('saveFlow', {nodes, edges, name: flowName}, (error) => {
-                if (error) {
-                    console.log(error.reason);
-                } else {
-                    setFlowName(''); // Clear the flow name input field after a successful save
-                    fetchFlows(); // Fetch the flows again to update the dropdown
-                }
-            });
-        }
-    };
-
-    const fetchFlows = () => {
-        Meteor.call('fetchFlows', (error, result) => {
-            if (error) {
-                console.error("Error fetching flows:", error);
-            } else {
-                setFetchedFlows(result);
-            }
-        });
-    };
-
-    const clearFlows = () => {
-        Meteor.call('clearFlows', (error) => {
-            if (error) {
-                console.error("Error clearing flows:", error);
-            } else {
-                console.log("Flows collection cleared");
-            }
-        });
-    };
-
-
-    const clearFlowchart = () => {
-        setNodes([]);
-        setEdges([]);
-    }
-
-    const loadFlow = (flow) => {
-        const newNodes = attachOnConnect(flow.nodes);
-        setNodes(newNodes);
-        setEdges(flow.edges);
-    };
-      
-
-    // Drag and drop state
-    const [draggedNodeType, setDraggedNodeType] = useState(null);
+    const {
+        reactFlowInstance,
+        updateReactFlowInstance,
+        clearFlowchart,
+        loadFlowchart,
+        nodes, 
+        updateNodes,
+        onNodesChange,
+        addNode,
+        edges, 
+        updateEdges, 
+        onEdgesChange,
+        handleConnect,
+        updateOutputNodes,
+    } = useLocalFlowData({nodes: [startingNode]});
 
     // Style states
-    const [sizes, setSizes] = useState(['60%', '50%']);
-    const [isDarkMode, setIsDarkMode] = useState(true);
+    const {sizes, handleSizeChange}= useResizeable();
 
-    const toggleDarkMode = () => {
-        setIsDarkMode(!isDarkMode);
-    };
+    // state for flows on db
+    const {
+        fetchedFlows, 
+        saveFlow, 
+        fetchFlows, 
+        clearFlows, 
+        flowName,
+        updateFlowName,
+    } = useRemoteFlowData(nodes, edges);
+      
+    // Drag and drop handling
+    const {handleDragStart, handleDragOver, handleDrop} = useDraggable(reactFlowInstance, addNode);
 
-    const handleSizeChange = (sizes) => {
-        setSizes(sizes);
-    };
-
-    const handleDragStart = (e, nodeType) => {
-        setDraggedNodeType(nodeType);
-        e.dataTransfer.setData('text/plain', nodeType);
-    };
-
-    const handleNodeChange = (id, value) => {
-        setNodes((ns) => {
-            const newNodes = ns.map((n) => 
-                n.id === id
-                    ? { ...n, data: { ...n.data, value } }
-                    : n
-            );
-            return newNodes;
-        });
-    };
-
-    // Function to add a new node
-    const addNode = (type, position) => {
-        const newNode = createNode(type, position, undefined, handleNodeChange);
-        setNodes((nodes) => [...nodes, newNode]);
-    };
-
-    const attachOnConnect = (nodes) => {
-        return nodes.map((node) => ({
-            ...node,
-            data: { ...node.data, onChange: (newValue) => { 
-                handleNodeChange(node.id, newValue) 
-            } },
-        }));
-    };
+    // Dark mode toggling
+    const {isDarkMode, toggleDarkMode} = useToggleable(true);
 
     useEffect(() => {
         fetchFlows();
@@ -129,50 +61,36 @@ const FlowchartCalculator = () => {
                 style={{ height: '100vh', width: '100%' }}
             >
                 <NodeButtons
-                    addNode={addNode}
                     handleDragStart={handleDragStart}
-                    isDarkMode={isDarkMode}
                     clearFlows={clearFlows}
                     clearFlowchart={clearFlowchart}
+                    isDarkMode={isDarkMode}
                 />
                 <FlowchartCanvas
+                    setReactFlowInstance={updateReactFlowInstance}
                     nodes={nodes}
                     edges={edges}
-                    setNodes={setNodes}
-                    setEdges={setEdges}
+                    onNodesChange={onNodesChange}
+                    onEdgesChange={onEdgesChange}
                     addNode={addNode}
-                    draggedNodeType={draggedNodeType}
-                    setDraggedNodeType={setDraggedNodeType}
+                    handleConnect={handleConnect}
+                    handleDragOver={handleDragOver}
+                    handleDrop={handleDrop}
+                    updateOutputNodes={updateOutputNodes}
                     isDarkMode={isDarkMode}
                 />
             </div>
             <JSONViewer 
                 nodes={nodes} 
                 edges={edges} 
-                isDarkMode={isDarkMode}
                 flows={fetchedFlows}
-                loadFlow={loadFlow}
-                saveFlow={saveFlow}
+                loadFlowchart={loadFlowchart}
                 onSave={saveFlow}
                 flowName={flowName}
-                setFlowName={setFlowName}
+                updateFlowName={updateFlowName}
+                isDarkMode={isDarkMode}
             >
-                <Switch
-                    checked={isDarkMode}
-                    onChange={toggleDarkMode}
-                    offColor='#f0f0f0'
-                    onColor='#606060'
-                    offHandleColor='#ccc'
-                    onHandleColor='#444'
-                    handleDiameter={30}
-                    uncheckedIcon={false}
-                    checkedIcon={false}
-                    boxShadow='0px 1px 5px rgba(0, 0, 0, 0.6)'
-                    // activeBoxShadow="0px 0px 1px 10px rgba(0, 0, 0, 0.2)"
-                    height={15}
-                    width={40}
-                    className='react-switch darkmode-switch'
-                />
+                <DarkModeSwitch isDarkMode={isDarkMode} toggleDarkMode={toggleDarkMode} />
             </JSONViewer>
         </SplitPane>
     );
