+++ seankwarren/flowchart-calculator/imports/ui/hooks/useInputField.jsx
@@ -1,0 +1,13 @@
+import React, { useState } from 'react'
+
+const useInputField = () => {
+    const [flowName, setFlowName] = useState('');
+
+    const updateFlowName = (e) => {
+        setFlowName(e.target.value);
+    };
+
+    return {flowName, updateFlowName}
+}
+
+export default useInputField
