+++ seankwarren/flowchart-calculator/imports/ui/hooks/useResizeable.jsx
@@ -1,0 +1,13 @@
+import React, { useState } from 'react'
+
+const useResizeable = () => {
+    const [sizes, setSizes] = useState(['60%', '50%']);
+
+    const handleSizeChange = (sizes) => {
+        setSizes(sizes);
+    };
+
+    return {sizes, handleSizeChange}
+}
+
+export default useResizeable
